refactor(dev): promote channel to top-level utility
since its used across `compiler/` and `devServer/`
