+++ packages/remix-dev/compiler/compileBrowser.ts
@@ -4,6 +4,7 @@ import * as esbuild from "esbuild";
 import { NodeModulesPolyfillPlugin } from "@esbuild-plugins/node-modules-polyfill";
 import { pnpPlugin as yarnPnpPlugin } from "@yarnpkg/esbuild-plugin-pnp";
 
+import { type WriteChannel } from "../channel";
 import { type RemixConfig } from "../config";
 import { createAssetsManifest, type AssetsManifest } from "./assets";
 import { getAppDependencies } from "./dependencies";
@@ -14,7 +15,6 @@ import { cssFilePlugin } from "./plugins/cssFilePlugin";
 import { emptyModulesPlugin } from "./plugins/emptyModulesPlugin";
 import { mdxPlugin } from "./plugins/mdx";
 import { urlImportsPlugin } from "./plugins/urlImportsPlugin";
-import { type WriteChannel } from "./utils/channel";
 import { writeFileSafe } from "./utils/fs";
 
 export type BrowserCompiler = {
