+++ packages/remix-dev/compiler/compilerServer.ts
@@ -4,6 +4,7 @@ import * as fse from "fs-extra";
 import { NodeModulesPolyfillPlugin } from "@esbuild-plugins/node-modules-polyfill";
 import { pnpPlugin as yarnPnpPlugin } from "@yarnpkg/esbuild-plugin-pnp";
 
+import { type ReadChannel } from "../channel";
 import { type RemixConfig } from "../config";
 import { type AssetsManifest } from "./assets";
 import { loaders } from "./loaders";
@@ -16,7 +17,6 @@ import { serverBareModulesPlugin } from "./plugins/serverBareModulesPlugin";
 import { serverEntryModulePlugin } from "./plugins/serverEntryModulePlugin";
 import { serverRouteModulesPlugin } from "./plugins/serverRouteModulesPlugin";
 import { urlImportsPlugin } from "./plugins/urlImportsPlugin";
-import { type ReadChannel } from "./utils/channel";
 
 export type ServerCompiler = {
   // produce ./build/index.js
